requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-ftp==3.6.0
apache-airflow-providers-http==4.6.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
argcomplete==3.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.0
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.2.0
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.18.1
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
google-re2==1.1
googleapis-common-protos==1.61.0
googlemaps==4.10.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kafka-python==2.0.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.4
mysql-connector-python==8.0.31
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
overrides==7.4.0
packaging==23.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prison==0.2.1
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==14.0.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pyspark==3.5.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.0
QtPy==2.4.1
ratelim==0.1.6
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rich-argparse==1.4.0
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.9
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0
